pyproject.toml
[project]
name = "doibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ryang Sohn", email = "loop.infinitely@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "requests>=2.32.5",
]

[project.scripts]
doibot = "doibot:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "dpytest>=0.7.0",
    "pytest>=9.0.1",
]
